made seperate component for sort option
