+++ src/components/HomeFeedView.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import RecipeTile from './RecipeTile.jsx'
+import RecipeTile from './RecipeTile.jsx';
+import SortOption from './SortOption.jsx';
 import downArrow from '../../dist/resources/homeView/down-arrow.png';
 import upArrow from '../../dist/resources/homeView/up-arrow.png';
 import { API_ADDR } from '../config';
@@ -11,76 +12,76 @@ import imageFour from '../../dist/resources/homeView/4.png';
 
 
 const HomeFeedView = ({captureNavigation, captureRecipeId, favorites, captureFavorites}) => {
-  const [sortDisplay, updateSortDisplay] = useState(false);
-  const [sortOption, updateSortOption] = useState('mostPopular');
+  // const [sortDisplay, updateSortDisplay] = useState(false);
+  // const [sortOption, updateSortOption] = useState('mostPopular');
   const [sortedData, updateSortedData] = useState([]);
   const [randomBackground, updaterandomBackground] = useState('');
 
 
-  const handleSortDisplay = () => {
-    let sortElement = document.getElementById('sortDropDown');
-    if (sortDisplay === false) {
-      updateSortDisplay(true);
-      sortElement.style.display = 'block';
-    } else {
-      updateSortDisplay(false);
-      sortElement.style.display = 'none';
-    }
-  }
+  // const handleSortDisplay = () => {
+  //   let sortElement = document.getElementById('sortDropDown');
+  //   if (sortDisplay === false) {
+  //     updateSortDisplay(true);
+  //     sortElement.style.display = 'block';
+  //   } else {
+  //     updateSortDisplay(false);
+  //     sortElement.style.display = 'none';
+  //   }
+  // }
 
-  const handleSortOption = (e) => {
-    let selectedSortOption = document.getElementById(sortOption);
-    selectedSortOption.style.color = 'black';
-    selectedSortOption.style.backgroundColor = '#f9f9f9';
+  // const handleSortOption = (e) => {
+  //   let selectedSortOption = document.getElementById(sortOption);
+  //   selectedSortOption.style.color = 'black';
+  //   selectedSortOption.style.backgroundColor = '#f9f9f9';
 
-    updateSortOption(e.target.id);
-    let sortElement = document.getElementById('sortDropDown');
-    sortElement.style.display = 'none';
-    handleSortDisplay();
-  }
+  //   updateSortOption(e.target.id);
+  //   let sortElement = document.getElementById('sortDropDown');
+  //   sortElement.style.display = 'none';
+  //   handleSortDisplay();
+  // }
 
-  const handleMostPopularRecipe = () => {
-    axios.get(`${API_ADDR}/recipes?sort=likes&direction=desc`)
-      .then ((res) => {
-        updateSortedData(res.data.rows);
-      })
-  }
+  // const handleMostPopularRecipe = () => {
+  //   axios.get(`${API_ADDR}/recipes?sort=likes&direction=desc`)
+  //     .then ((res) => {
+  //       updateSortedData(res.data.rows);
+  //     })
+  // }
 
-  const handleMostExpensive = () => {
-    axios.get(`${API_ADDR}/recipes?sort=price&direction=desc`)
-      .then ((res) => {
-        updateSortedData(res.data.rows);
-      })
-  }
+  // const handleMostExpensive = () => {
+  //   axios.get(`${API_ADDR}/recipes?sort=price&direction=desc`)
+  //     .then ((res) => {
+  //       updateSortedData(res.data.rows);
+  //     })
+  // }
 
-  const handleCheapest = () => {
-    axios.get(`${API_ADDR}/recipes?sort=price&direction=asc`)
-      .then ((res) => {
-        updateSortedData(res.data.rows);
-      })
-  }
+  // const handleCheapest = () => {
+  //   axios.get(`${API_ADDR}/recipes?sort=price&direction=asc`)
+  //     .then ((res) => {
+  //       updateSortedData(res.data.rows);
+  //     })
+  // }
 
-  useEffect (() => {
-    if (sortOption === 'mostPopular') {
-      let selectedSortOption = document.getElementById(sortOption);
-      selectedSortOption.style.color = 'white';
-      selectedSortOption.style.backgroundColor = '#E86D4D';
+  // useEffect (() => {
+  //   if (sortOption === 'mostPopular') {
+  //     let selectedSortOption = document.getElementById(sortOption);
+  //     selectedSortOption.style.color = 'white';
+  //     selectedSortOption.style.backgroundColor = '#E86D4D';
 
-      handleMostPopularRecipe();
-    } else if (sortOption === 'highPrice') {
-      let selectedSortOption = document.getElementById(sortOption);
-      selectedSortOption.style.color = 'white';
-      selectedSortOption.style.backgroundColor = '#E86D4D';
+  //     handleMostPopularRecipe();
+  //   } else if (sortOption === 'highPrice') {
+  //     let selectedSortOption = document.getElementById(sortOption);
+  //     selectedSortOption.style.color = 'white';
+  //     selectedSortOption.style.backgroundColor = '#E86D4D';
 
-      handleMostExpensive();
-    } else if (sortOption === 'lowPrice') {
-      let selectedSortOption = document.getElementById(sortOption);
-      selectedSortOption.style.color = 'white';
-      selectedSortOption.style.backgroundColor = '#E86D4D';
+  //     handleMostExpensive();
+  //   } else if (sortOption === 'lowPrice') {
+  //     let selectedSortOption = document.getElementById(sortOption);
+  //     selectedSortOption.style.color = 'white';
+  //     selectedSortOption.style.backgroundColor = '#E86D4D';
 
-      handleCheapest();
-    }
-  }, [sortOption])
+  //     handleCheapest();
+  //   }
+  // }, [sortOption])
 
   useEffect (() => {
     updaterandomBackground([imageOne,imageTwo,imageThree,imageFour][Math.floor(Math.random()*4)]);
@@ -98,7 +99,8 @@ const HomeFeedView = ({captureNavigation, captureRecipeId, favorites, captureFav
       <div id="homeBottomView">
         <div id="titleHomeBottomView">
           <div id="title">Recommended Recipes</div>
-          <div id="sort">
+          <SortOption captureNavigation={captureNavigation} captureRecipeId={captureRecipeId} favorites={favorites} captureFavorites={captureFavorites}/>
+          {/* <div id="sort">
             <div id="sortByText">Sort:</div>
             {sortOption === 'mostPopular' ? <h4 id="sortDescription" onClick={handleSortDisplay}>Most Popular</h4> : ''}
             {sortOption === 'highPrice' ? <h4 id="sortDescription" onClick={handleSortDisplay}>Price: High-Low</h4> : ''}
@@ -111,15 +113,15 @@ const HomeFeedView = ({captureNavigation, captureRecipeId, favorites, captureFav
                 <p onClick={handleSortOption} id="highPrice">Price: High-Low</p>
                 <p onClick={handleSortOption} id="lowPrice">Price: Low-High</p>
               </div>
-          </div>
+          </div> */}
         </div>
-        <div id="receipeBox">
+        {/* <div id="receipeBox">
           {sortedData[0] ? (
             sortedData.map((recipe) => {
               return <RecipeTile key={recipe.id} recipe={recipe} captureNavigation={captureNavigation} captureRecipeId={captureRecipeId} favorites={favorites} captureFavorites={captureFavorites}/>
             })
           ) : ''}
-        </div>
+        </div> */}
       </div>
     </div>
   );
